Finalize zpool property, update man page
The previous commit implemented the persistent L2ARC as a zpool property.
When changing the zpool property "l2cache_persistent" the zpool had to be
exported and re-imported for the L2ARC to be rebuilt. This commit rebuilds
the L2ARC asynchronously  when l2cache_persistent is set to "yes", without
requiring the pool to be exported and re-imported.

Signed-off-by: George Amanakis <gamanakis@gmail.com>