Fix the calculation in l2arc_log_blk_overhead()
The overhead for the log blocks is calculated as follows:
Divide the amount of bytes to be written or evicted with
SPA_MINBLOCKSIZE. This gives the amount of headers or log entries to be
written. Multiply by the size of one log entry (128 bytes). To this
amount add the overhead for the headers of the log blocks. Divide the
log entries to be written by the amount of entries in a log block and
multiply with the length of the header of a log block.

Signed-off-by: George Amanakis <gamanakis@gmail.com>