Improvements on persistent L2ARC
Multiple improvements on code readability and corrected comments.

Issue #10224 raised concern whether we are correctly accounting for
the size of L2ARC log blocks. We mitigate this by
1) Implementing a new arcstat "l2_log_blk_asize" which reflects how
   l2arc_evict() and l2arc_log_blk_commit() account for the size of
   L2ARC log blocks. This is also written to the header of the cache
   device as dh_lb_asize. zdb can compare dh_lb_asize to the sum of
   all aligned sizes as reported by l2arc_rebuild(). This provides
   valuable information as to whether l2arc_evict() evicts fewer log
   blocks than it should. If those two values are not equal zdb exits
   with an error. For this feature to work correctly there should be
   no active writes on the cache device. This is also employed in the
   tests of persistent L2ARC. Since we modify the structure of the
   cache device header, this change breaks backward compatibility in
   terms of persistent L2ARC.
2) In l2arc_evict() we make sure that the tail of the list of log
   block pointers points to a log block whose payload starts after
   the evict hand. In other words, the evict hand lies between the
   writing hand and the start of the payload of the log block pointed
   by the tail of the list.

Make the first test in persistent L2ARC use `zdb -lll` to increase
coverage in `zdb.c`.

In l2arc_rebuild_vdev() if the amount of committed log entries in a log
block is 0 and the device header is valid we update the device header.
This will facilitate trimming of the whole device in this case when
TRIM for L2ARC is implemented.

Signed-off-by: George Amanakis <gamanakis@gmail.com>